# autotest-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Avtolider API
  description: Uzbekistan UBDD Driving License Exam Preparation Platform
  version: 1.0.0
  contact:
    name: Avtolider Team
servers:
  - url: http://localhost:5000/api/v1
    description: Development
  - url: https://api.avtolider.uz/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LocalizedText:
      type: object
      required: [uz, uzLatin, ru]
      properties:
        uz:
          type: string
          description: Uzbek Cyrillic
        uzLatin:
          type: string
          description: Uzbek Latin
        ru:
          type: string
          description: Russian
    PaginationMeta:
      type: object
      required: [page, pageSize, totalCount, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalCount:
          type: integer
        totalPages:
          type: integer
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'

    # ─── Auth ───────────────────────────────────
    SendOtpRequest:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
          pattern: '^998[0-9]{9}$'
    SendOtpResponse:
      type: object
      required: [expiresInSeconds, retryAfterSeconds]
      properties:
        expiresInSeconds:
          type: integer
        retryAfterSeconds:
          type: integer
    VerifyOtpRequest:
      type: object
      required: [phoneNumber, code]
      properties:
        phoneNumber:
          type: string
          pattern: '^998[0-9]{9}$'
        code:
          type: string
          minLength: 6
          maxLength: 6
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, expiresAt, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: integer
          description: Unix timestamp
        user:
          $ref: '#/components/schemas/UserDto'
    TelegramLoginRequest:
      type: object
      required: [id, firstName, authDate, hash]
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        authDate:
          type: integer
        hash:
          type: string
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    UserDto:
      type: object
      required: [id, phoneNumber, role, subscriptionStatus, preferredLanguage, createdAt]
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        subscriptionStatus:
          type: string
          enum: [none, active, expired, cancelled]
        preferredLanguage:
          type: string
          enum: [uz, uzLatin, ru]
        createdAt:
          type: string
          format: date-time

    # ─── Tickets ────────────────────────────────
    TicketDto:
      type: object
      required: [ticketNumber, questionCount]
      properties:
        ticketNumber:
          type: integer
        questionCount:
          type: integer

    # ─── Categories ─────────────────────────────
...
    AnswerOptionDto:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
    QuestionDto:
      type: object
      required: [id, text, difficulty, categoryId, categoryName, ticketNumber, answerOptions]
      properties:
        id:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
        difficulty:
          type: integer
          enum: [1, 2, 3]
        categoryId:
          type: string
          format: uuid
        categoryName:
          $ref: '#/components/schemas/LocalizedText'
        ticketNumber:
          type: integer
        answerOptions:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOptionDto'

    # ─── Exams ──────────────────────────────────
    StartExamRequest:
      type: object
      required: [licenseCategory]
      properties:
        templateId:
          type: string
          format: uuid
          nullable: true
        licenseCategory:
          type: string
          enum: [AB, CD]
    ExamQuestionDto:
      type: object
      required: [id, questionId, order, text, answerOptions]
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        order:
          type: integer
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
        answerOptions:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOptionDto'
    ExamSessionDto:
      type: object
      required: [id, status, totalQuestions, passingScore, timeLimitMinutes, mode, questions]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [inProgress, completed, expired]
        totalQuestions:
          type: integer
        passingScore:
          type: integer
        timeLimitMinutes:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Null for marathon mode (no timer)
        mode:
          type: string
          enum: [exam, ticket, marathon]
        ticketNumber:
          type: integer
          nullable: true
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ExamQuestionDto'
    SubmitAnswerRequest:
      type: object
      required: [sessionQuestionId, selectedAnswerId]
      properties:
        sessionQuestionId:
          type: string
          format: uuid
        selectedAnswerId:
          type: string
          format: uuid
        timeSpentSeconds:
          type: integer
          nullable: true
    SubmitAnswerResponse:
      type: object
      required: [recorded, answeredCount, totalCount]
      properties:
        recorded:
          type: boolean
        answeredCount:
          type: integer
        totalCount:
          type: integer
    ExamResultQuestionDto:
      type: object
      required: [questionId, text, selectedAnswerId, correctAnswerId, isCorrect]
      properties:
        questionId:
          type: string
          format: uuid
        text:
          $ref: '#/components/schemas/LocalizedText'
        imageUrl:
          type: string
          nullable: true
        selectedAnswerId:
          type: string
          format: uuid
          nullable: true
        correctAnswerId:
          type: string
          format: uuid
        isCorrect:
          type: boolean
        explanation:
          $ref: '#/components/schemas/LocalizedText'
        timeSpentSeconds:
          type: integer
          nullable: true
    ExamResultDto:
      type: object
      required: [examId, score, correctAnswers, totalQuestions, passingScore, passed, timeTakenSeconds, completedAt, questions]
      properties:
        examId:
          type: string
          format: uuid
        score:
          type: integer
        correctAnswers:
          type: integer
        totalQuestions:
          type: integer
        passingScore:
          type: integer
        passed:
          type: boolean
        timeTakenSeconds:
          type: integer
        completedAt:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: '#/components/schemas/ExamResultQuestionDto'
    ExamHistoryItemDto:
      type: object
      required: [examId, score, totalQuestions, passed, completedAt, timeTakenSeconds]
      properties:
        examId:
          type: string
          format: uuid
        score:
          type: integer
        totalQuestions:
          type: integer
        passed:
          type: boolean
        completedAt:
          type: string
          format: date-time
        timeTakenSeconds:
          type: integer

    # ─── Practice ───────────────────────────────
...
paths:
...
  /auth/send-otp:
    post:
      tags: [Auth]
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SendOtpResponse'
        '429':
          description: Rate limited
  /auth/verify-otp:
    post:
      tags: [Auth]
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/telegram-login:
    post:
      tags: [Auth]
      operationId: telegramLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramLoginRequest'
      responses:
        '200':
          description: Telegram login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/refresh-token:
    post:
      tags: [Auth]
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserDto'

  # ═══════════ CATEGORIES ═══════════
...
  /questions:
    get:
      tags: [Questions]
      operationId: getQuestions
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: difficulty
          in: query
          schema:
            type: integer
            enum: [1, 2, 3]
        - name: ticketNumber
          in: query
          schema:
            type: integer
        - name: lang
          in: query
          schema:
            type: string
            enum: [uz, uzLatin, ru]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated questions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/QuestionDto'
                          meta:
                            $ref: '#/components/schemas/PaginationMeta'

  # ═══════════ TICKETS ═══════════
  /questions/tickets:
    get:
      tags: [Questions]
      operationId: getTicketsList
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tickets with question counts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketDto'
  /questions/tickets/{number}:
    get:
      tags: [Questions]
      operationId: getQuestionsByTicket
      security:
        - bearerAuth: []
      parameters:
        - name: number
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Questions for a specific ticket
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/QuestionDto'

  # ═══════════ EXAMS ═══════════
  /exams/start:
    post:
      tags: [Exams]
      operationId: startExam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExamRequest'
      responses:
        '200':
          description: Exam started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamSessionDto'
  /exams/start-ticket:
    post:
      tags: [Exams]
      operationId: startTicketExam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticketNumber]
              properties:
                ticketNumber:
                  type: integer
      responses:
        '200':
          description: Ticket exam started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamSessionDto'
  /exams/start-marathon:
    post:
      tags: [Exams]
      operationId: startMarathon
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Marathon started (all questions, no timer)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamSessionDto'
  /exams/{examId}/answer:
    post:
      tags: [Exams]
      operationId: submitExamAnswer
      security:
        - bearerAuth: []
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Answer recorded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmitAnswerResponse'
  /exams/{examId}/complete:
    post:
      tags: [Exams]
      operationId: completeExam
      security:
        - bearerAuth: []
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamResultDto'
  /exams/{examId}:
    get:
      tags: [Exams]
      operationId: getExamSession
      security:
        - bearerAuth: []
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam session
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ExamSessionDto'
  /exams/history:
    get:
      tags: [Exams]
      operationId: getExamHistory
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Exam history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: '#/components/schemas/ExamHistoryItemDto'
                          meta:
                            $ref: '#/components/schemas/PaginationMeta'

  # ═══════════ PRACTICE ═══════════
...